+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Mock-up test file.
+- Helpers test file.
+
+### Changed
+- Router tests.
+
 ## [0.5.1] - 2017-12-21
 ### Fixed
 - `Cancel` not canceling middleware stacks.
@@ -68,6 +76,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
+[Unreleased]: https://github.com/gbrlsnchs/httpmux/compare/v0.5.1...HEAD
 [0.5.1]: https://github.com/gbrlsnchs/httpmux/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.3.0...v0.4.0